// src/enums/title.enum.ts
export enum title {
    home = "Home",
    job = "Job",
    explore = "Explore",
    blog = "Blog",
    contact = "Contact",
    dashboard = "Dashboard",
    candidateDashboard = "Candidate Dashboard",
    employerDashboard = "Employer Dashboard",
    candidate = "Candidate",
    employer = "Employer",
    company = "Company",
    logIn = "Log in",
    register = "Register",
    demandingCategory = "Most Demanding Categories."
}